requirements.txt
streamlit>=1.45.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.9.2
loguru>=0.7.2
langchain>=0.3.3
langgraph>=0.2.38
humanize>=4.11.0
streamlit-tags>=1.2.8
streamlit-scrollable-textbox>=0.0.3
tiktoken>=0.8.0
streamlit-monaco>=0.1.3
langgraph-checkpoint-sqlite>=2.0.0
langchain-community>=0.3.3
langchain-openai>=0.2.3
langchain-anthropic>=0.2.3
langchain-ollama>=0.2.0
langchain-google-genai>=2.0.1
langchain-groq>=0.2.1
langchain_mistralai>=0.2.1
langchain_deepseek>=0.1.3
tomli>=2.0.2
groq>=0.12.0
python-dotenv>=1.0.1
httpx[socks]>=0.27.0
nest-asyncio>=1.6.0
content-core>=1.0.2
ai-prompter>=0.3
esperanto>=2.0.4
langchain-google-vertexai>=2.0.10
surrealdb>=1.0.4
surreal-commands>=1.0.13
podcast-creator>=0.2.6
langgraph-checkpoint-postgres==2.0.23
psycopg==3.2.9
psycopg-pool==3.2.6
orjson==3.11.1
asyncpg==0.30.0